lms_backend/pyproject.toml
[project]
name = "lms-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosmtpd==1.4.6",
    "django>=5.2",
    "django-extensions==4.1",
    "django-rest-knox>=5.0.2",
    "django-stubs==5.1.3",
    "djangorestframework==3.16.0",
    "djoser>=2.3.1",
    "mysqlclient==2.2.7",
    "social-auth-app-django>=5.4.3",
    "sshtunnel>=0.4.0",
]
